% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{methods}
\name{any,Diff-method}
\alias{any,Diff-method}
\title{Determine if Diff Object Has Differences}
\usage{
\S4method{any}{Diff}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a \code{Diff} object}

\item{...}{unused, for compatibility with generic}

\item{na.rm}{unused, for compatibility with generic}
}
\value{
TRUE if there are differences, FALSE if not, FALSE with warning if
  there are no differences but objects are not \code{\link{all.equal}}
}
\description{
Determine if Diff Object Has Differences
}
\examples{
any(diffChr(letters, letters))
any(diffChr(letters, letters[-c(1, 5, 8)]))
}
