% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{as.character,diffrMyersMbaSes-method}
\alias{as.character,diffrMyersMbaSes-method}
\title{Generate a character representation of Shortest Edit Sequence}
\usage{
\S4method{as.character}{diffrMyersMbaSes}(x, ...)
}
\arguments{
\item{x}{S4 object of class \code{diffrMyersMbaSes}}

\item{...}{unused}
}
\value{
character vector
}
\description{
Generate a character representation of Shortest Edit Sequence
}
\seealso{
\code{\link{diff_ses}}
}

