% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summmary.R
\docType{methods}
\name{as.character,DiffSummary-method}
\alias{as.character,DiffSummary-method}
\title{Generate Character Representation of DiffSummary Object}
\usage{
\S4method{as.character}{DiffSummary}(x, ...)
}
\arguments{
\item{x}{a \code{DiffSummary} object}

\item{...}{not used, for compatibility with generic}
}
\value{
the summary as a character vector intended to be \code{cat}ed to
  terminal
}
\description{
Generate Character Representation of DiffSummary Object
}
\examples{
as.character(
  summary(diffChr(letters, letters[-c(5, 15)], format="raw", pager="off"))
)
}

