% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{as.character,MyersMbaSes-method}
\alias{as.character,MyersMbaSes-method}
\title{Generate a character representation of Shortest Edit Sequence}
\usage{
\S4method{as.character}{MyersMbaSes}(x, ...)
}
\arguments{
\item{x}{S4 object of class \code{MyersMbaSes}}

\item{...}{unused}
}
\value{
character vector
}
\description{
Generate a character representation of Shortest Edit Sequence
}
\seealso{
\code{\link{ses}}
}
\keyword{internal}
