% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{auto_context}
\alias{auto_context}
\alias{auto_line_limit}
\title{Configure Automatic Parameter Calculations}
\usage{
auto_context(min = getOption("diffobj.context.auto.min"),
  max = getOption("diffobj.context.auto.max"))
}
\arguments{
\item{min}{integer(1L), positive, set to zero to allow any context}

\item{max}{integer(1L), set to negative to allow any context}

\item{limit}{integer(2L), see documentation for \code{line.limit} parameter
for \code{link{diff_obj}}}

\item{pager.lines}{integer(1L), number of lines above which output is
displayed with \code{\link{file.show}} instead of \code{cat}ed directly to
screen.  Set to negative to always use pager, zero to never use pager.
Defaults to \code{\link{console_lines} * 1.2}}

\item{less.flags}{character(1L), what flags to set with the \code{LESS}
system environment variable.  This is only relevant if your system is
configured to use \code{less} as the pager.  You should only provide the
flag letters (e.g. \code{"RX"}, not \code{"-RX"}).  Defaults to \code{"R"}
if it appears that \code{less} is the system pager and the \code{use.ansi}
\code{\link{diff_settings}}
The system variable is only modified for the duration of the evaluation and
is reset / unset afterwards.}
}
\value{
S4 object containing configuration parameters, for use as the
  \code{context} or \code{line_limit} parameters value in
  \code{\link{diff_obj}} and related functions
}
\description{
Helper functions to help define parameters for selecting an appropriate
\code{context} and \code{line.limit} values.
}

