% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{console_lines}
\alias{console_lines}
\title{Attempt to Compute Console Height in Text Lines}
\usage{
console_lines()
}
\value{
integer(1L)
}
\description{
Returns the value of the \code{LINES} system variable if it is reasonable,
48 otherwise.
}
\examples{
console_lines()
}

