% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{class}
\name{Diff-class}
\alias{Diff-class}
\title{Diff Result Object}
\description{
Return value for the \code{\link[=diffPrint]{diff*}} methods.  Has
\code{show}, \code{as.character}, \code{summmary}, \code{[}, \code{head},
\code{tail}, and \code{any} methods.
}
