% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{diff_myers}
\alias{diff_myers}
\title{Diff two character vectors}
\usage{
diff_myers(a, b, max.diffs = 0L)
}
\arguments{
\item{a}{character}

\item{b}{character}

\item{max.diffs}{integer(1L) how many differences before giving up; set to
zero to allow as many as there are}
}
\value{
list
}
\description{
Implementation of Myer's Diff algorithm with linear space refinement
originally implemented by Mike B. Allen as part of
\href{libmba}{http://www.ioplex.com/~miallen/libmba/}
version 0.9.1.  This implementation is a heavily modified version of the
original C code and is not compatible with the \code{libmba} library.
The C code is simplified by using fixed size arrays instead of variable
ones for tracking the longest reaching paths and for recording the shortest
edit scripts.  Additionally all error handling and memory allocation calls
have been moved to the internal R functions designed to handle those things.
A failover result is provided in the case where max diffs allowed is
exceeded.  Ability to provide custom comparison functions is removed.
}
\keyword{internal}

