% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_obj}
\alias{diff_obj}
\title{Diff Objects}
\usage{
diff_obj(target, current, mode = getOption("diffobj.mode"),
  context = getOption("diffobj.context"), etc = NULL, ...)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object being compared to \code{target}}

\item{mode}{character(1L), one of:
\itemize{
  \item unified: diff mode used by \code{git diff}, and the default here
  \item sidebyside: line up the differences side by side
  \item context: show the target and current hunks in their entirety; this
    mode takes up a lot of screen space but makes it easier to see what the
    objects actually look like
}}

\item{context}{integer(1L) how many lines of context are shown on either side
of differences, set to \code{-1L} to allow as many as there are.  Set to
\dQuote{auto} (default) to display as many as 10 lines or as few as 1
depending on whether total screen lines fit within \code{line.limit} (see
\code{\link{etc}}), or alternatively pass the return value of
\code{\link{auto_context}} to fine tune the parameters of the auto context
calculation.}

\item{etc}{diffObjSettings S4 object as generated by \code{\link{etc}},
allows control of parameters such as whether to truncate output, use a
pager, ignore white space differences, color output, etc.  See
documentation for \code{\link{etc}} for more details.  If NULL (default)
equivalent to passing \code{etc()}}

\item{...}{additional arguments to pass on to \code{print}, \code{str}, etc.}
}
\value{
a \code{\link{diffObjDiff}} object; this object has a \code{show}
  method that will display the diff to screen
}
\description{
Compare either the \code{print}ed or \code{str} screen representation of
R objects depending on which is estimated to produce the most useful
diff.  The selection process tries to minimize screen lines while maximizing
differences shown subject to display constraints.  The decision algorithm is
likely to evolve over time, so do not rely on this function making a
a particular selection under specific circumstances.  Instead, use
\code{\link{diff_print}} or \code{\link{diff_str}} if you require one or the
other output.
}
\seealso{
\code{\link{etc}} for more detailed control of diff settings,
  \code{\link{diff_print}}, \code{\link{diff_str}},
  \code{\link{diff_chr}} to compare character vectors directly,
  \code{\link{diff_deparse}} to compare deparsed objects
}

