% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{diff_ses}
\alias{diff_ses}
\title{Produce Shortest Edit Script}
\usage{
diff_ses(a, b)
}
\arguments{
\item{a}{character}

\item{b}{character}
}
\value{
character
}
\description{
Intended primarily for debugging or for other applications that understand
that particular format.  See \href{GNU diff docs}{http://www.gnu.org/software/diffutils/manual/diffutils.html#Detailed-Normal}
for how to interpret the symbols.
}

