% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_str}
\alias{diff_str}
\title{Diff Object Structures}
\usage{
diff_str(target, current, mode = getOption("diffobj.mode"),
  context = getOption("diffobj.context"), etc = etc(), ...)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object being compared to \code{target}}

\item{mode}{character(1L), one of:
\itemize{
  \item unified: diff mode used by \code{git diff}, and the default here
  \item sidebyside: line up the differences side by side
  \item context: show the target and current hunks in their entirety; this
    mode takes up a lot of screen space but makes it easier to see what the
    objects actually look like
}}

\item{context}{integer(1L) how many lines of context are shown on either side
of differences, set to \code{-1L} to allow as many as there are.  Set to
\dQuote{auto} (default) to display as many as 10 lines or as few as 1
depending on whether total screen lines fit within \code{line.limit} (see
\code{\link{etc}}), or alternatively pass the return value of
\code{\link{auto_context}} to fine tune the parameters of the auto context
calculation.}

\item{etc}{diffObjSettings S4 object as generated by \code{\link{etc}},
allows control of parameters such as whether to truncate output, use a
pager, ignore white space differences, color output, etc.  See
documentation for \code{\link{etc}} for more details.}

\item{...}{additional arguments to pass on to \code{print}, \code{str}, etc.}
}
\value{
a \code{\link{diffObjDiff}} object; this object has a \code{show}
  method that will display the diff to screen
}
\description{
Compares the \code{str} output of \code{target} and \code{current}.  If
the \code{max.level} parameter to \code{str} is left unspecified, will
attempt to find the largest \code{max.level} that fits within
\code{line.limit} (see \code{\link{etc}}) and shows at least one difference.
}
\details{
Due to the seemingly inconsistent nature of \code{max.level} when used with
objects with nested attributes, and also due to the relative slowness of
\code{str}, this function simulates the effect of \code{max.level} by hiding
nested lines instead of repeatedly calling \code{str} with varying values of
\code{max.level}.
}
\seealso{
\code{\link{diff_obj}} for details on output and diff algorithm,
  \code{\link{etc}} for more detailed control of diff settings,
  \code{\link{diff_print}},
  \code{\link{diff_chr}} to compare character vectors directly,
  \code{\link{diff_deparse}} to compare deparsed objects
}

