% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{diff_style}
\alias{diff_style}
\title{Specify Styles to Use in Diffs}
\usage{
diff_style(line.ins = identity, line.del = identity,
  line.match = identity, line = identity, gutter = identity,
  gutter.ins = identity, gutter.del = identity, word.ins = identity,
  word.del = identity, banner = identity, banner.ins = identity,
  banner.del = identity, hunk.header = identity, meta = identity)
}
\arguments{
\item{line.ins}{style for every line that exists in \code{current} and not in
\code{target}}

\item{line.del}{style for every line that exists in \code{target} and not in
\code{current}}

\item{line}{outermost style for every line}

\item{gutter}{style for every gutter element}

\item{gutter.ins}{style for every gutter element corresponding to an insertion}

\item{gutter.del}{style for every gutter element corresponding to a deletion}

\item{word.ins}{style for every word that exists in \code{current} and not in
\code{target} when a line is eligble for word diffing}

\item{word.del}{style for every word that exists in \code{target} and not in
\code{current} when a line is eligble for word diffing}

\item{banner}{style around entire banner}

\item{banner.ins}{style around insertion portion of the banner}

\item{banner.del}{style around deletion portion of the banner}

\item{hunk.header}{style around each hunk header}

\item{meta}{style around meta information lines}
}
\description{
Styles should be functions that require a single character argument and
return a character vector the same length as the input.  The intended use
case is to pass \code{crayon} functions such as \code{\link{crayon::red}},
although you may pass any function of your liking that behaves as described.
}
\details{
Every component that is styled differently is exposed through this function.
the \code{line} style will wrap every line (note that each row in a
side-by-side diff contains two lines).  Each \code{line.*} style will wrap
around all \code{gutter} and \code{word} styles.
}

