% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diffCsv}
\alias{diffCsv}
\title{Diff CSV Files}
\usage{
diffCsv(target, current, ...)
}
\arguments{
\item{target}{character(1L) or file connection with read capability;
if character should point to a CSV file}

\item{current}{like \code{target}}

\item{...}{additional arguments to pass on to \code{print}, \code{str}, etc.}
}
\value{
a \code{\link{Diff}} object; this object has a \code{show}
  method that will display the diff to screen
}
\description{
Reads CSV files with \code{\link{read.csv}} and passes the resulting data
frames onto \code{\link{diffPrint}}.  \code{...} arguments are passed to
both \code{read.csv} and \code{print}.  To the extent you wish to use
different \code{...} arguments for each of those functions you will need to
\code{read.csv} the files and pass them to \code{diffPrint} yourself.
}
\examples{
iris.2 <- iris
iris.2$Sepal.Length[5] <- 99
f1 <- tempfile()
f2 <- tempfile()
write.csv(iris, f1, row.names=FALSE)
write.csv(iris.2, f2, row.names=FALSE)
diffCsv(f1, f2)
unlink(c(f1, f2))
}
\seealso{
\code{\link{diffPrint}} for details on the \code{diff*} functions,
  \code{\link{diffObj}}, \code{link{diffStr}},
  \code{\link{diffChr}} to compare character vectors directly
}

