% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffobj-package}
\alias{diffobj-package}
\title{Compare R Objects with a Text Diff}
\description{
Highlights differences between R objects in a familiar and intuitive way.
This is similar to \code{\link{tools::Rdiff}} except the diff is computed
directly on R objects instead of text files, does not rely on the system
\code{diff} utility, and provides alternate display modes.
}

