% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{diffobj_css}
\alias{diffobj_css}
\title{Return Location of Default CSS File}
\usage{
diffobj_css()
}
\value{
path to the default CSS file
}
\description{
Used as the value for \code{getOption("diffobj.html.css")}.  Note that the
contents of this file rather than the file address are used when
constructing HTML output.
}

