% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R, R/core.R, R/tochar.R
\docType{methods}
\name{diffobj_s4method_doc}
\alias{any,diffObjDiffDiffs-method}
\alias{as.character,diffObjDiff-method}
\alias{as.character,diffObjDiffHtml-method}
\alias{diffobj_s4method_doc}
\alias{show,diffObjMyersMbaSes-method}
\alias{summary,diffObjMyersMbaSes-method}
\title{Dummy Doc File for S4 Methods with Existing Generics}
\usage{
\S4method{any}{diffObjDiffDiffs}(x, ..., na.rm = FALSE)

\S4method{show}{diffObjMyersMbaSes}(object)

\S4method{summary}{diffObjMyersMbaSes}(object, with.match = FALSE, ...)

\S4method{as.character}{diffObjDiff}(x, ...)

\S4method{as.character}{diffObjDiffHtml}(x, ...)
}
\description{
Dummy Doc File for S4 Methods with Existing Generics
}
\keyword{internal}

