% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{diffobj_set_def_opts}
\alias{diffobj_set_def_opts}
\title{Set All diffobj Options to Defaults}
\usage{
diffobj_set_def_opts()
}
\value{
list for use with \code{options} that contains values of
  \code{diffob} options before they were forced to defaults
}
\description{
Used primarily for testing to ensure all options are set to default values.
}
\examples{
\dontrun{
  diffobj_set_def_opts()
}
}
