% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{diffobj_settings}
\alias{dfos}
\alias{diffobj_settings}
\title{Generate a \code{diffobj} Setting Object}
\usage{
diffobj_settings(line.limit = getOption("diffobj.line.limit"),
  hunk.limit = getOption("diffobj.hunk.limit"), pager = pager_settings(),
  ignore.white.space = getOption("diffobj.ignore.white.space"),
  use.ansi = getOption("diffobj.use.ansi"), disp.width = getOption("width"),
  silent = getOption("diffobj.silent"),
  max.diffs = getOption("diffobj.max.diffs"),
  max.diffs.in.hunk = getOption("diffobj.max.diffs.in.hunk"),
  max.diffs.wrap = getOption("diffobj.max.diffs.wrap"),
  tab.stops = getOption("diffobj.tab.stops"), frame = parent.frame(),
  tar.banner = NULL, cur.banner = NULL)
}
\arguments{
\item{hunk.limit}{integer(2L) how many sections of differences to show.
Behaves similarly to the \code{line.limit} parameter for
\code{link{diff_obj}} except it only supports numeric input (defaults to
-1L, i.e. unlimited)}

\item{use.ansi}{TRUE or FALSE, whether to use ANSI escape sequences to color
differences (TRUE by default if we detect that your terminal supports it)}

\item{disp.width}{integer(1L) number of display columns to take up; note that
in \dQuote{sidebyside} mode the effective display width is half this number}

\item{silent}{TRUE or FALSE, whether to suppress display the diff (FALSE by
default).}

\item{max.diffs}{integer(1L), number of differences after which we abandon
the \code{O(n^2)} diff algorithm in favor of a linear one.  Set to
\code{-1L} to always stick to the original algorithm (defaults to 10000L).}

\item{max.diffs.in.hunk}{integer(1L), like \code{max.diffs}, but used when
computing word diffs within hunks.  Used independently for each hunk.  Set
to zero to turn off in-hunk word diff.}

\item{max.diffs.wrap}{integer(1L), like \code{max.diffs}, but used when
computing word diffs on atomic vectors (defaults to 10000L), see
\dQuote{Atomic Vectors} for \code{\link{diff_print}}.}

\item{tab.stops}{integer, what tab stops to use when converting hard tabs to
spaces.  If not integer will be coerced to integer.}

\item{frame}{environment the evaluation frame for the \code{print/show/str},
calls, allows user to ensure correct methods are used, not used by
\code{\link{diff_chr}} or \code{\link{diff_deparse}}.}

\item{tar.banner}{character(1L) or NULL, used to clarify the symbology of the
diff output (see the \dQuote{Output} section in the docs), if NULL will be
inferred from \code{target} and \code{current} expressions.}

\item{cur.banner}{character(1L) like \code{tar.banner}, but for \code{current}}

\item{white.space}{TRUE or FALSE, whether to consider differences in
horizontal whitespace (i.e. spaces and tabs) as differences (defaults to
FALSE)}
}
\value{
a \code{diffObjSettings} S4 object
}
\description{
Used for more fine grained control on the \code{diff_obj} family of functions.
\code{dfos} is just shorthand.
}

