% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{class}
\name{diffObjStyleFuns-class}
\alias{diffObjStyleBasic}
\alias{diffObjStyleBasic-class}
\alias{diffObjStyleBasicYB}
\alias{diffObjStyleBasicYB-class}
\alias{diffObjStyleDark}
\alias{diffObjStyleDark-class}
\alias{diffObjStyleDarkYB}
\alias{diffObjStyleDarkYB-class}
\alias{diffObjStyleFuns}
\alias{diffObjStyleFuns-class}
\alias{diffObjStyleHtml}
\alias{diffObjStyleHtml-class}
\alias{diffObjStyleHtmlYB}
\alias{diffObjStyleHtmlYB-class}
\alias{diffObjStyleLight}
\alias{diffObjStyleLight-class}
\alias{diffObjStyleLightYB}
\alias{diffObjStyleLightYB-class}
\title{Customize Appearance of Diff}
\arguments{
\item{line}{function}

\item{line.ins}{function}

\item{line.del}{function}

\item{line.match}{function}

\item{text}{function}

\item{text.ins}{function}

\item{text.del}{function}

\item{text.match}{function}

\item{gutter}{function}

\item{gutter.ins}{function}

\item{gutter.del}{function}

\item{gutter.match}{function}

\item{hunk.header}{function to format each hunk header with}

\item{banner.ins}{function to format insertion banner}

\item{banner.del}{function to format deletion banner}

\item{banner}{function to format entire banner}

\item{meta}{function format meta information lines}

\item{gutter.ins.txt}{character(1L) text to use as visual cue to indicate
whether a diff line is an insertion, defaults to \dQuote{> }}

\item{gutter.ins.txt.ctd}{character(1L) if a diff line is wrapped, the
visual cue shifts to this character to indicate wrapping occured}

\item{gutter.del.txt}{character(1L) see \code{gutter.ins.txt} above}

\item{gutter.del.txt.ctd}{character(1L) see \code{gutter.ins.txt.ctd} above}

\item{gutter.match.txt}{character(1L) see \code{gutter.ins.txt} above}

\item{gutter.match.txt.ctd}{character(1L) see \code{gutter.ins.txt.ctd} above}
}
\value{
diffObjStyle S4 object
}
\description{
Most of the customization is done by specifying functions that operate on
character vectors and return a modified character vector of the same length.
The intended use case is to pass \code{crayon} functions such as
\code{\link{crayon::red}}, although you may pass any function of your liking
that behaves as described.
}
\details{
The visual representation of the diff has many nested components.  The
functions you specify here will be applied with the function corresponding to
the innermost element applied first.  A schematic of the various component
that represent an inserted line follows:
\preformatted{+- line ----------------------------------------+
|+- line.ins ----------------------------------+|
||+- gutter ---------+ +- text ---------------+||
|||+- gutter.ins ---+| |+- text.ins ---------+|||
||||                || ||      +- word.ins -+||||
|||| gutter.ins.txt || || DIFF | TEXT HERE  |||||
||||                || ||      +------------+||||
|||+----------------+| |+--------------------+|||
||+------------------+ +----------------------+||
|+---------------------------------------------+|
+-----------------------------------------------+
}
A similar model applies to deleted and matching lines.  The boxes represent
functions.  \code{gutter.ins.txt} represents the text to use in the gutter
and is not a function. \code{DIFF TEXT HERE} is text from the objects being
diffed, with the portion that has different words inside the \code{word.ins}
box provided word diff is enabled, and is obviously not a function either.

Most of the functions defined here default to \code{\link{identity}}, but
you are given the flexibility to fully format the diff.
}
\note{
in \dQuote{sidebyside} there are two lines per row of text, one showing
  deletions and one showing additions.
}

