\name{diffr-package}
\alias{diffr-package}
\alias{diffr}
\docType{package}
\title{
\packageTitle{diffr}
}
\description{
\packageDescription{diffr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diffr}
\packageIndices{diffr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{diffr}

Maintainer: \packageMaintainer{diffr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
