% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diffr}
\alias{diffr}
\title{Diff two character vectors}
\usage{
diffr(a, b)
}
\arguments{
\item{a}{character}

\item{b}{character}
}
\value{
list
}
\description{
Implementation of Myer's diff borrowed from Mike B. Allen
}

