% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{methods}
\name{diffr_s4method_doc}
\alias{any,diffrDiff-method}
\alias{as.character,diffrDiff-method}
\alias{diffr_s4method_doc}
\title{Dummy Doc File for S4 Methods with Existing Generics}
\usage{
\S4method{any}{diffrDiff}(x, ..., na.rm = FALSE)

\S4method{as.character}{diffrDiff}(x, context, width, ...)
}
\description{
Dummy Doc File for S4 Methods with Existing Generics
}
\keyword{internal}

