% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{methods}
\name{dimnames,PaletteOfStyles-method}
\alias{dimnames,PaletteOfStyles-method}
\title{Retrieve Dimnames for PaletteOfStyles Objects}
\usage{
\S4method{dimnames}{PaletteOfStyles}(x)
}
\arguments{
\item{x}{a \code{\link{PaletteOfStyles}} object}
}
\value{
list the dimension names
dimnames(PaletteOfStyles())
}
\description{
Retrieve Dimnames for PaletteOfStyles Objects
}

