% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalizer.R, R/s4.R, R/summmary.R
\name{finalizeHtml}
\alias{finalizeHtml}
\alias{finalizeHtml,ANY-method}
\alias{finalizeHtml,Diff-method}
\alias{finalizeHtml,DiffSummary-method}
\title{Finalizing Methods for HTML Output}
\usage{
finalizeHtml(x, ...)

\S4method{finalizeHtml}{ANY}(x, x.chr, js, ...)

\S4method{finalizeHtml}{Diff}(x, x.chr, ...)

\S4method{finalizeHtml}{DiffSummary}(x, x.chr, ...)
}
\arguments{
\item{x}{object to finalize}

\item{...}{arguments to pass on to methods}

\item{x.chr}{character text representation of \code{x}, typically generated
with the \code{as.character} method for \code{x}}

\item{js}{character javascript code to append to HTML representation}
}
\description{
Used as the \code{finalizer} slot to \code{\link{StyleHtml}} objects to wrap
character output prior to output to device.  Used primarily by styles that
output to HTML to properly configure HTML page structure, including injecting
JS, CSS, etc..
}
