% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalizer.R
\name{finalizer}
\alias{finalizer}
\title{Finalizing Methods}
\usage{
finalizer(x, ...)
}
\description{
Use as the \code{finalizer} slot to \code{\link{Style}} objects to wrap
character output prior to output to device.  Used primarily by styles that
output to HTML to properly configure HTML page structure.
}

