% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{gdo}
\alias{gdo}
\title{Shorthand Function for Accessing diffobj Options}
\usage{
gdo(x)
}
\arguments{
\item{x}{character(1L) name off \code{diffobj} option to retrieve, without
the \dQuote{diffobj.} prefix}
}
\description{
\code{gdo(x)} is equivalent to \code{getOption(sprintf("diffobj.\%s", x))}.
}
\examples{
gdo("format")
}

