% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\name{printGuideLines}
\alias{chrGuideLines}
\alias{deparseGuideLines}
\alias{printGuideLines}
\alias{strGuideLines}
\title{Generic Methods to Implement Flexible Guide Line Computations}
\usage{
printGuideLines(obj, obj.as.chr)

strGuideLines(obj, obj.as.chr)

chrGuideLines(obj, obj.as.chr)

deparseGuideLines(obj, obj.as.chr)
}
\description{
These functions are most useful
}

