% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiff.R
\name{has_Rdiff}
\alias{has_Rdiff}
\title{Attempt to Detect Whether 'diff' Utility Available}
\usage{
has_Rdiff()
}
\value{
TRUE or FALSE
}
\description{
Checks whether \code{tools::Rdiff} issues a warning and if it does assumes
this is because the diff utility is not available.
}

