% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{nchar_html}
\alias{nchar_html}
\title{Count Text Characters in HTML}
\usage{
nchar_html(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{unused for compatibility with internal use}
}
\value{
integer(length(x)) with number of characters of each element
}
\description{
Very simple implementation that will fail if there are any \dQuote{>} in the
HTML that are not closing tags, and assumes that HTML entities are all one
character wide.  Also, spaces are counted as one width each because the
HTML output is intended to be displayed inside \code{<PRE>} tags.
}
\examples{
nchar_html("<a href='http:www.domain.com'>hello</a>")
}
