% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{pager_is_less}
\alias{pager_is_less}
\title{Check Whether System Has less as Pager}
\usage{
pager_is_less()
}
\value{
TRUE or FALSE
}
\description{
If \code{getOption(pager)} is set to the default value, checks whether
\code{Sys.getenv("PAGER")} appears to be \code{less} by trying to run the
pager with the \dQuote{version} flag and confirming that the file name is
also \code{less}.  If \code{getOption(pager)} is not the default value, then
checks whether it points to the \code{less} program.
}
\details{
In some systems \code{more} is just a copy of \code{less}, but because in
those cases the binary does not appear to always respond to the
\code{$LESS} environment variable it is treated as if it is not \code{less}.
}
\examples{
pager_is_less()
}

