% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{pager_settings}
\alias{pager_settings}
\title{Control Under What Circumstances Output is Displayed Through Pager}
\usage{
pager_settings(mode = getOption("diffobj.pager.mode"),
  threshold = getOption("diffobj.pager.threshold"),
  less.flags = getOption("diffobj.less.flags"))
}
\arguments{
\item{mode}{character(1L) one of \itemize{
  \item threshold: use pager if output has more lines than \code{threshold}
  \item always: always use pager
  \item never: never use pager
}}

\item{threshold}{integer(1L) if in \code{mode} "threshold", number of lines
of output that triggers the use of the pager; negative values lead to
using \code{\link{console_lines}} + 1}

\item{less.flags}{character(1L), what flags to set with the \code{LESS}
system environment variable.  This is only relevant if your system is
configured to use \code{less} as the pager.  You should only provide the
flag letters (e.g. \code{"RX"}, not \code{"-RX"}).  Defaults to \code{"R"}
if it appears that \code{less} is the system pager and the \code{use.ansi}
setting is TRUE.  The system variable is only modified for the duration of
the evaluation and is reset / unset afterwards.}
}
\value{
S4 object for use as the \code{pager} parameter to
  \code{link{diffobj_settings}}
}
\description{
Pager is invoked via \code{\link{file.show}}.  This function is intended for
use with the \code{pager} parameter for \code{\link{diffobj_settings}}.
}

