% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{class}
\name{PaletteOfStyles-class}
\alias{PaletteOfStyles}
\alias{PaletteOfStyles-class}
\title{Class for Tracking Default Styles by Style Type}
\description{
Provides a mechanism for specifying a style based on the style properties
along dimensions of format, brightness, and color.  This allows a user to
request a style that meets a certain description (e.g. a \dQuote{light}
scheme in \dQuote{ansi256} format), without having to provide a specific
\code{\link{Style}} object.
}
\section{Dimensions}{


There are three general orthogonal dimensions of styles that can be used when
rendering diffs: the type of format, the \dQuote{brightness} of the output,
and whether the colors used are distinguishable if you assume reds and greens
are not distinguishable.  Defaults for the intersections each of these
dimensions are encoded as a three dimensional list.  This list is just an
atomic vector of type \dQuote{list} with a length 3 \code{dim} attribute.

The array/list dimensions are:
\itemize{
  \item format: the format type, typically \dQuote{raw}, \dQuote{ansi8},
    \dQuote{ansi256}, or \dQuote{html}
  \item brightness: whether the colors are bright or not, which allows user to
    chose a scheme that is compatible with their console, typically:
    \dQuote{light}, \dQuote{dark}, \dQuote{normal}
  \item color.mode: \dQuote{rgb} for full color or \dQuote{yb} for
    dichromats (yb stands for Yellow Blue).
}
}

\section{Structural Details}{


The array/list is stored in the \code{data} slot of
\code{PaletteOfStyles} objects.  Subsetting methods are provided so you
may operate directly on the S4 object as you would on a regular array.

The array/list must be fully populated with objects that are or extend
\code{Style}.  There is no explicit check that the objects in the list
comply with the descriptions implied by their coordinates, although the
default object provided by the package does comply for the most part.  One
check that is carried out is that any element that has a \dQuote{html}
value in the \code{format} dimension extends \code{StyleHtml}.

Every cell in the list must be populated.  If there is a particular
combination of coordinates that does not have a corresponding defined style
a reasonable substitution must be provided.  For example, this package
only defines \dQuote{light} HTML styles, so it simply uses that style for
all the possible \code{brightness} values.

While the list may only have the three dimensions described, you can add
values to the dimensions provided the values described above are the first
ones in each of their corresponding dimensions.  For example, if you wanted
to allow for styles that would render in \code{grid} graphics, you could
genarate a default list with \dQuote{"grid"} value appended to the values of
the \code{format} dimension.
}
\examples{
## Create a new style based on existing style by changing
## gutter symbols
my.style <- StyleAnsi256LightRgb()
my.style@text@gutter.ins <- "+"
my.style@text@gutter.del <- "-"
## Generate the default style object palette, and replace
## the ansi256 / light / rgb style with our modified one
defs <- PaletteOfStyles()
defs["ansi256", "light", "rgb"] <- list(my.style) # note `list()`
## If so desired, set our new style palette as the default
## one; could also pass directly as argument to `diff*` funs
\dontrun{
options(diffobj.style.defaults) <- defs
}
}

