% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.diff_myers_mba}
\alias{print.diff_myers_mba}
\title{Print Method for Shortest Edit Path}
\usage{
\method{print}{diff_myers_mba}(x, with.match = FALSE, ...)
}
\arguments{
\item{x}{the \code{diff_myers_mba} object to display}

\item{with.match}{logical(1L) whether to show what text the edit command
refers to}

\item{...}{forwarded to the data frame print method used to actually display
the data}
}
\value{
whatever the data frame print method returns
}
\description{
Displays the data required to generate the shortest edit path for comparison
between two strings.
}

