% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowhead.R
\name{printTrim}
\alias{printTrim}
\title{Trim Method for Printed Objects}
\usage{
printTrim(obj, obj.as.chr)
}
\arguments{
\item{obj}{the object}

\item{obj.as.chr}{charcter the \code{print}ed representation of the object}
}
\value{
a \code{length(obj.as.chr) * 2} integer matrix with the start (first
  column and end (second column) character positions of the sub string to
  run diffs on.
}
\description{
Trim Method for Printed Objects
}
\note{
\code{obj.as.chr} will be post \code{strip_hz_control}
}

