% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{Rdiff_obj}
\alias{Rdiff_obj}
\title{a \code{tools::Rdiff} Between R Objects}
\usage{
Rdiff_obj(from, to, ...)
}
\arguments{
\item{from}{an R object (see details)}

\item{to}{another R object (see details)}

\item{...}{passed on to \code{Rdiff}}
}
\value{
whatever \code{Rdiff} returns
}
\description{
Just a wrapper that saves the \code{print} / \code{show} representation of an
object to a temp file and then runs \code{tools::Rdiff} on them.  For
each of \code{from}, \code{to}, will check if they are 1 length character
vectors referencing an RDS file, and will use the contents of that RDS file
as the object to compare.
}
\seealso{
\code{tools::Rdiff}
}

