% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ses}
\alias{ses}
\title{Shortest Edit Script}
\usage{
ses(a, b)
}
\arguments{
\item{a}{character}

\item{b}{character}
}
\value{
character
}
\description{
Computes shortest edit script to convert \code{a} into \code{b} by removing
elements from \code{a} and adding elements from \code{b}.  Intended primarily
for debugging or for other applications that understand that particular
format.  See \href{GNU diff docs}{http://www.gnu.org/software/diffutils/manual/diffutils.html#Detailed-Normal}
for how to interpret the symbols.
}
\examples{
ses(letters[1:3], letters[2:4])
}

