% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summmary.R
\docType{methods}
\name{show,DiffSummary-method}
\alias{show,DiffSummary-method}
\title{Display DiffSummary Objects}
\usage{
\S4method{show}{DiffSummary}(object)
}
\arguments{
\item{object}{a \code{DiffSummary} object}
}
\value{
NULL, invisbly
}
\description{
Display DiffSummary Objects
}

