% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{methods}
\name{show,PaletteOfStyles-method}
\alias{show,PaletteOfStyles-method}
\title{Display a PaletteOfStyles}
\usage{
\S4method{show}{PaletteOfStyles}(object)
}
\arguments{
\item{object}{a \code{\link{PaletteOfStyles}} object}
}
\value{
NULL, invisibly
}
\description{
Display a PaletteOfStyles
}
