% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{methods}
\name{show,Style-method}
\alias{show,Style-method}
\alias{show,StyleHtml-method}
\title{Show Method for Style Objects}
\usage{
\S4method{show}{Style}(object)

\S4method{show}{StyleHtml}(object)
}
\arguments{
\item{object}{a \code{Style} S4 object}
}
\value{
NULL, invisibly
}
\description{
Display a small sample diff with the Style object styles applied.  For
ANSI light and dark styles, will also temporarily set the background and
foreground colors to ensure they are compatible with the style, even though
this is not done in normal output (i.e. if you intend on using a
\dQuote{light} style, you should set your terminal background color to be
light or expect sub-optimal rendering).
}
\examples{
show(StyleAnsi256LightYb())  # assumes ANSI colors supported
}

