% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{class}
\name{StyleSummary-class}
\alias{StyleSummary}
\alias{StyleSummary-class}
\title{Styling Information for Summaries}
\arguments{
\item{container}{function applied to entire summary}

\item{body}{function applied to everything except the actual map portion of
the summary}

\item{map}{function applied to the map portion of the summary}
}
\description{
Styling Information for Summaries
}

