% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{class}
\name{StyleSummary-class}
\alias{StyleSummary}
\alias{StyleSummary-class}
\alias{StyleSummaryHtml}
\alias{StyleSummaryHtml-class}
\title{Styling Information for Summaries}
\description{
Styling Information for Summaries
}
\section{Slots}{

\describe{
\item{\code{container}}{function applied to entire summary}

\item{\code{body}}{function applied to everything except the actual map portion of
the summary}

\item{\code{detail}}{function applied to section showing how many deletions /
insertions, etc. occurred}

\item{\code{map}}{function applied to the map portion of the summary}
}}

