% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{[,Diff,numeric,missing,missing-method}
\alias{[,Diff,numeric,missing,missing-method}
\title{Subsetting Methods for Diff Objects}
\usage{
\S4method{[}{Diff,numeric,missing,missing}(x, i)
}
\arguments{
\item{x}{\code{Diff} object}

\item{i}{subsetting index, must be numeric}
}
\description{
Methods to subset the character representation of the diff output.  The
subsetting bears no link to the line numbers in the diffs, only to the actual
displayed diff.
}
\details{
\code{[} only supports numeric indices, and returns without error if you
specify out of bound indices.  If you apply multiple subsetting methods they
will be applied in the following order irrespective of what order you 
actually specify them in: \code{[}, then \code{head}, then \code{tail}.
If you use the same subsetting method multiple times on the same object,
the last call will define the outcome.

These methods are implemented by storing the chosen indices in the 
\code{Diff} object and using them to subset the \code{as.character} output.
This mechanism explains the seemingly odd behavior documented above.
}

