% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{summary,MyersMbaSes-method}
\alias{summary,MyersMbaSes-method}
\title{Summary Method for Shortest Edit Path}
\usage{
\S4method{summary}{MyersMbaSes}(object, with.match = FALSE, ...)
}
\arguments{
\item{object}{the \code{diff_myers} object to display}

\item{with.match}{logical(1L) whether to show what text the edit command
refers to}

\item{...}{forwarded to the data frame print method used to actually display
the data}
}
\value{
whatever the data frame print method returns
}
\description{
Displays the data required to generate the shortest edit path for comparison
between two strings.
}
\keyword{internal}
