% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{summary,PaletteOfStyles-method}
\alias{summary,PaletteOfStyles-method}
\title{Display a Summarized Version of a PaletteOfStyles}
\usage{
\S4method{summary}{PaletteOfStyles}(object, ...)
}
\arguments{
\item{object}{a \code{\link{PaletteOfStyles}} object}

\item{...}{unused, for compatibility with generic}
}
\value{
character representation showing classes and/or objects in
  PaletteOfStyles
summary(PaletteOfStyles())
}
\description{
Display a Summarized Version of a PaletteOfStyles
}
