% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{tag_f}
\alias{div_f}
\alias{span_f}
\alias{tag_f}
\title{Make Functions That Wrap Text in HTML Tags}
\usage{
tag_f(tag, class = character(), style = character())

div_f(class = character(), style = character())

span_f(class = character(), style = character())
}
\arguments{
\item{character}{class the CSS class(es)}

\item{named}{character style inline styles, where the name is the CSS
property and the value the value.}
}
\value{
a function that accepts a character parameter.  If applied, each
  element in the character vector will be wrapped in the div tags
}
\description{
See examples
}
\note{
inputs are assumed to be valid class names or CSS styles.
}
\examples{
## Assuming class 'ex1' has CSS styles defined elsewhere
tag_f("div", "ex1")(LETTERS[1:5])
## Use convenience function, and add some inline styles
div_f("ex2", c(color="green", `font-family`="arial"))(LETTERS[1:5])
}
\keyword{export}

