% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{view_or_browse}
\alias{view_or_browse}
\title{Invoke IDE Viewer if Available}
\usage{
view_or_browse(url)
}
\arguments{
\item{url}{character(1L) a location containing a file to display}
}
\value{
the return vaue of \code{getOption("viewer")} if it is a function, or
  of \code{\link{browseURL}}
}
\description{
Use \code{getOption("viewer")} to view HTML output if it is available as 
per \link{RStudio}{https://support.rstudio.com/hc/en-us/articles/202133558-Extending-RStudio-with-the-Viewer-Pane}. Fallback to \code{\link{browseURL}} 
if not available.
}

