% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{auto_context}
\alias{auto_context}
\title{Configure Automatic Contextz Calculation}
\usage{
auto_context(min = getOption("diffobj.context.auto.min"),
  max = getOption("diffobj.context.auto.max"))
}
\arguments{
\item{min}{integer(1L), positive, set to zero to allow any context}

\item{max}{integer(1L), set to negative to allow any context}
}
\value{
S4 object containing configuration parameters, for use as the
  \code{context} or parameter value in \code{\link{diff_obj}} and related
  functions
}
\description{
Helper functions to help define parameters for selecting an appropriate
\code{context} value.
}

