% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_obj}
\alias{diff_obj}
\title{Diff Objects}
\usage{
diff_obj(target, current, mode = getOption("diffobj.mode"),
  context = getOption("diffobj.context"), etc = NULL, ...)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object being compared to \code{target}}

\item{mode}{character(1L), one of:
\itemize{
  \item unified: diff mode used by \code{git diff}, and the default here
  \item sidebyside: line up the differences side by side
  \item context: show the target and current hunks in their entirety; this
    mode takes up a lot of screen space but makes it easier to see what the
    objects actually look like
}}

\item{context}{integer(1L) how many lines of context are shown on either side
of differences, set to \code{-1L} to allow as many as there are.  Set to
\dQuote{auto} (default) to display as many as 10 lines or as few as 1
depending on whether total screen lines fit within \code{line.limit} (see
\code{\link{etc}}), or alternatively pass the return value of
\code{\link{auto_context}} to fine tune the parameters of the auto context
calculation.}

\item{etc}{diffObjSettings S4 object as generated by \code{\link{etc}},
allows control of parameters such as whether to truncate output, use a
pager, ignore white space differences, color output, etc.  See
documentation for \code{\link{etc}} for more details.  If NULL (default)
equivalent to passing \code{etc()}}

\item{...}{additional arguments to pass on to \code{print}, \code{str}, etc.}
}
\value{
a \code{\link{diffObjDiff}} object; this object has a \code{show}
  method that will display the diff to screen
}
\description{
Compare either the \code{print}ed or \code{str} screen representation of
R objects depending on which is estimated to produce the most useful
diff.  The selection process tries to minimize screen lines while maximizing
differences shown subject to display constraints.  The decision algorithm is
likely to evolve over time, so do not rely on this function making a
a particular selection under specific circumstances.  Instead, use
\code{\link{diff_print}} or \code{\link{diff_str}} if you require one or the
other output.
}
\section{Output}{


The result of the diff provides the information necessary to transform the
\code{target} object into the \code{current} object.  This involves deletions
from and additions to \code{target}.  The deletions and additions are done
linewise.  Each deleted line will have \code{- } prepended to it, and each
added line will have \code{+ } prepended to it.  If your terminal supports
ANSI escape sequences the additions and deletions will be color coded.

The first lines of output clarify the coding convention by showing the
\code{target} object with the deletion symbology, and the \code{current}
object with the addition symbology.  After these lines you will see the
first and possibly only hunk header.  The format will be \code{@ x,y z,w @}
where \code{x} and \code{z} indicate the starting line of the text in
the \code{target} and \code{current} objects that is shown after the hunk
header.  \code{y} and \code{w} indicate how many lines from each of those
objects are being shown.

In addition to the primary line diff, hunks are themselves word-diffed within
each hunk to help quickly identify small differences.  Just keep in mind that
the \code{+-} symbols always relate to the original line diff.  The
word-diff is indicated only by the ANSI escape sequence styling and will not
be visible if your terminal does not support them or if you disable them.

The output format used here is loosely based on the \code{git diff} format.
}

\section{Diff Algorithm}{


The diff algorithm is Myer's solution to the shortest edit script /
longest common sequence problem with the Hirschberg linear space refinement
as described in:
\cite{
E. Myers, \dQuote{An O(ND) Difference Algorithm and Its Variations},
Algorithmica 1, 2 (1986), 251-266.
\url{http://www.cs.arizona.edu/people/gene/PAPERS/diff.ps}
}
and should be the same algorithm used by GNU diff.  The implementation
used here is an adaptation of Michael B. Allen's diff program from the
\href{
   http://www.ioplex.com/~miallen/libmba/dl/libmba-0.9.1.tar.gz
}{\code{libmba}} \code{C} library.

This algorithm scales with the \bold{square} of the number of differences
between compared objects so is most effective when comparing objects
that are mostly similar.
}
\seealso{
\code{\link{etc}} for more detailed control of diff settings,
  \code{\link{diff_print}}, \code{\link{diff_str}},
  \code{\link{diff_chr}} to compare character vectors directly,
  \code{\link{diff_deparse}} to compare deparsed objects
}

