% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_obj}
\alias{diff_chr}
\alias{diff_obj}
\alias{diff_print,}
\alias{diff_str,}
\title{Show Diffs Between the Screen Display Versions of Two Objects}
\usage{
diff_obj(target, current, context = NULL, white.space = FALSE)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object being compared to \code{target}}

\item{context}{2 length integer vector representing how many lines of context
are shown on either side of differences.  The first value is the maximum
before we start trimming output.  The second value is the maximum to be
shown before we start trimming.  We will always attempt to show as much as
\code{2 * context + 1} lines of output so context may not be centered if
objects display as less than \code{2 * context + 1} lines.}

\item{white.space}{TRUE or FALSE, whether to consider differences in
horizontal whitespace (i.e. spaces and tabs) as differences (defaults to
FALSE)}

\item{max.level}{integer(1L) up to how many levels to try running \code{str};
\code{str} is run repeatedly starting with \code{max.level=1} and then
increasing \code{max.level} until we fill the context or a difference
appears or the \code{max.level} specified here is reached.  If the value is
reached then will let \code{str} run with \code{max.level} unspecified.
This is designed to produce the most compact screen output possible that
shows the differences between objects, though obviously it comes at a
performance cost; set to 0 to disable}
}
\value{
character, invisibly, the text representation of the diff
}
\description{
Designed to highlight at a glance the \bold{display} differences between
two objects.  Lack of visual differences is not guarantee that the objects
are the same.  These functions are designed to help you quickly understand
the nature of differences between objects when they are known to be different
(e.g. not \code{identical} or \code{all.equal}).  The diff algorithms are far
from perfect and in some cases will likely make seemingly odd choices on what
to highlight as being different.
}
\details{
These functions focus on the first display difference between two objects.
If you want to see the full object diff try \code{\link{Rdiff_obj}}.

\itemize{
  \item \code{diff_print} shows the differences in the \code{print} or
    \code{show} screen output of the two objects
  \item \code{diff_str} shows the differences in the \code{str} screen output
    of the two objects; will show as many recursive levels as possible so
    long as context lines are not exceeded, and if they are, as few as
    possible to show at least one error (see \code{max.level})
  \item \code{diff_obj} picks between \code{diff_print} and \code{diff_str}
    depending on which one it thinks will provide the most useful diff
}
}
\note{
: differences shown or reported by these functions may not be the
  totality of the differences between objects since display methods may not
  display all differences.  This is particularly true when using \code{str}
  for comparisons with \code{max.level} since differences inside unexpanded
  recursive levels will not be shown at all.
}

