% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_obj}
\alias{diff_obj}
\title{Diff Objects}
\usage{
diff_obj(target, current, mode = gdo("mode"), context = gdo("context"),
  line.limit = gdo("line.limit"), format = gdo("format"),
  brightness = gdo("brightness"), color.mode = gdo("color.mode"),
  pager = gdo("pager"), ignore.white.space = gdo("ignore.white.space"),
  max.diffs = gdo("max.diffs"), align.threshold = gdo("align.threshold"),
  disp.width = gdo("disp.width"), hunk.limit = gdo("hunk.limit"),
  convert.hz.white.space = gdo("convert.hz.white.space"),
  tab.stops = gdo("tab.stops"), style = gdo("style"),
  palette.of.styles = StylePalette(), tar.banner = NULL,
  cur.banner = NULL, frame = parent.frame(), ...)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object being compared to \code{target}}

\item{mode}{character(1L), one of:
\itemize{
  \item \dQuote{unified}: diff mode used by \code{git diff}, and the
    default here
  \item \dQuote{sidebyside}: line up the differences side by side
  \item \dQuote{context}: show the target and current hunks in their
    entirety; this mode takes up a lot of screen space but makes it easier
    to see what the objects actually look like
}}

\item{context}{integer(1L) how many lines of context are shown on either side
of differences, set to \code{-1L} to allow as many as there are.  Set to
\dQuote{auto} (default) to display as many as 10 lines or as few as 1
depending on whether total screen lines fit within \code{line.limit} (see
\code{\link{etc}}), or alternatively pass the return value of
\code{\link{auto_context}} to fine tune the parameters of the auto context
calculation.}

\item{line.limit}{integer(2L) or integer(1L), if length 1 how many lines of
output to show, where \code{-1} means no limit.  If length 2, the first
value indicates the threshold of screen lines to begin truncating output,
and the second the number of lines to truncate to, which should be fewer
than the threshold.}

\item{ignore.white.space}{TRUE or FALSE, whether to consider differences in
horizontal whitespace (i.e. spaces and tabs) as differences (defaults to
FALSE)}

\item{max.diffs}{integer(1L), number of differences after which we abandon
the \code{O(n^2)} diff algorithm in favor of a linear one.  Set to
\code{-1L} to always stick to the original algorithm (defaults to 10000L).}

\item{align.threshold}{numeric(1L) between 0 and 1, proportion of
characters in a line of \code{target} that must be matched in a line of
\code{current} in the same hunk for those lines to be paired up when
displayed (defaults to 0.25).}

\item{disp.width}{integer(1L) number of display columns to take up; note that
in \dQuote{sidebyside} mode the effective display width is half this
number (defaults to \code{getOption("width")}.}

\item{hunk.limit}{integer(2L) or integer (1L), how many diff hunks to show.
Behaves similarly to the \code{line.limit}.  How many hunks are in a
particular diff is a function of how many differences, and also how much
\code{context} is used since context can cause two hunks to bleed into
each other and become one.}

\item{tab.stops}{integer, what tab stops to use when converting hard tabs to
spaces.  If not integer will be coerced to integer (defaults to 8L).}

\item{tar.banner}{character(1L) or NULL, text to display ahead of the diff
section representing the target output.  If NULL will be
inferred from \code{target} and \code{current} expressions.}

\item{cur.banner}{character(1L) like \code{tar.banner}, but for
\code{current}}

\item{frame}{environment the evaluation frame for the \code{print/show/str},
calls, allows user to ensure correct methods are used, not used by
\code{\link{diff_chr}} or \code{\link{diff_deparse}}.}

\item{...}{additional arguments to pass on to \code{print}, \code{str}, etc.}
}
\value{
a \code{\link{Diff}} object; this object has a \code{show}
  method that will display the diff to screen
}
\description{
Compare either the \code{print}ed or \code{str} screen representation of
R objects depending on which is estimated to produce the most useful
diff.  The selection process tries to minimize screen lines while maximizing
differences shown subject to display constraints.  The decision algorithm is
likely to evolve over time, so do not rely on this function making a
a particular selection under specific circumstances.  Instead, use
\code{\link{diff_print}} or \code{\link{diff_str}} if you require one or the
other output.
}
\seealso{
\code{\link{etc}} for more detailed control of diff settings,
  \code{\link{diff_print}}, \code{\link{diff_str}},
  \code{\link{diff_chr}} to compare character vectors directly,
  \code{\link{diff_deparse}} to compare deparsed objects
}

