% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffobj-package}
\alias{diffobj-package}
\title{Compare R Objects with a Diff}
\description{
Diffs provide a clear and intuitive accounting of the differences between two
objects because they present the objects in usual way except modified to
highlight the differences.  This package contains a collection of functions
for carrying out and displaying colorized diffs on R objects for display in
a terminal or in a browser.  This is inspired by
\code{\link[=Rdiff]{tools::Rdiff}}, except the diff is
computed directly on R objects instead of text files, the diff computation
is independent of the system \code{diff} utility, and the diff display is
colorized and optimized to handle common R objects.
}

