% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffobj-package}
\alias{diffobj-package}
\title{Compare R Objects with a Text Diff}
\description{
Compare R objects by computing a \code{diff} on their text representations
(e.g. the \code{print}ed console output).  This is similar to
\code{\link{tools::Rdiff}} except the diff is computed directly on R objects
instead of text files and does not rely on the system \code{diff} utility.
}

