% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffobj-package}
\alias{diffobj-package}
\title{Compare R Objects with a Diff}
\description{
Colorized diffs to quickly identify _and understand_ differences between R
objects.  See `vignette("diffobj")` for details.
}

