% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffobj-package}
\alias{diffobj-package}
\title{Diffs for R Objects}
\description{
Generate a colorized diff of two R objects for an intuitive visualization of
their differences.  See `vignette(package="diffobj", "diffobj")` for details.
}
