% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R, R/core.R, R/tochar.R
\docType{methods}
\name{diffobj_s4method_doc}
\alias{any,DiffDiffs-method}
\alias{as.character,Diff-method}
\alias{diffobj_s4method_doc}
\alias{show,MyersMbaSes-method}
\alias{summary,MyersMbaSes-method}
\title{Dummy Doc File for S4 Methods with Existing Generics}
\usage{
\S4method{any}{DiffDiffs}(x, ..., na.rm = FALSE)

\S4method{show}{MyersMbaSes}(object)

\S4method{summary}{MyersMbaSes}(object, with.match = FALSE, ...)

\S4method{as.character}{Diff}(x, ...)
}
\description{
Dummy Doc File for S4 Methods with Existing Generics
}
\keyword{internal}

