% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R, R/core.R, R/tochar.R
\name{diffobj_s4method_doc}
\alias{diffobj_s4method_doc}
\alias{show,MyersMbaSes-method}
\alias{as.character,Diff-method}
\title{Dummy Doc File for S4 Methods with Existing Generics}
\usage{
\S4method{show}{MyersMbaSes}(object)

\S4method{as.character}{Diff}(x, ...)
}
\description{
Dummy Doc File for S4 Methods with Existing Generics
}
\keyword{internal}
