% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\docType{package}
\name{diffr-package}
\alias{diffr-package}
\title{Run Text Diffs on R Objects}
\description{
Implements tools similar to \code{\link{tools::Rdiff}} that operate directly
on R objects instead of R terminal output stored in \code{.out} files.  The
difference computations are based on the Myers algorithm with the linear
space modification as implemented by Mike B. Allen in \code{libmba-0.9.1}.
This is the same algorithm used by default by the GNU diff UNIX utility.
}

