% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{pager_is_less}
\alias{pager_is_less}
\title{Check Whether System Has less as Pager}
\usage{
pager_is_less()
}
\value{
TRUE or FALSE
}
\description{
Checks system \code{PAGER} variable and that \code{PAGER_PATH} is pointed
at \dQuote{R_HOME/bin/pager}.  This is an approximation and may return
false positives or negatives depending on your system.
}
\examples{
pager_is_less()
}

