% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{show,diffObjMyersMbaSes-method}
\alias{show,diffObjMyersMbaSes-method}
\title{Print Method for Shortest Edit Path}
\usage{
\S4method{show}{diffObjMyersMbaSes}(object)
}
\arguments{
\item{object}{object to display}
}
\value{
character the shortest edit path character representation, invisibly
}
\description{
Bare bones display of shortest edit path using GNU diff conventions
}

