% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{strip_hz_control}
\alias{strip_hz_control}
\title{Replace Horizontal Spacing Control Characters}
\usage{
strip_hz_control(txt, stops = 8L)
}
\arguments{
\item{txt}{character to covert}

\item{stops}{integer, what tab stops to use}
}
\description{
Removes tabs, newlines, and carriage returns and manipulates the text so that
it looks the renders the same as it did with those horizontal control
characters embedded.  This function is used when the
\code{convert.hz.white.space} parameter to the
\code{\link[=diffPrint]{diff*}} methods is active.  The term \dQuote{strip}
is a misnomer that remains for legacy reasons and lazyness.
}
\details{
This is an internal function with exposed documentation because it is
referenced in an external function's documentation.
}
\keyword{internal}

