% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{class}
\name{StyleText-class}
\alias{StyleText}
\alias{StyleText-class}
\title{Character Tokens Used in Diffs}
\arguments{
\item{gutter.insert}{character(1L) text to use as visual cue to indicate
whether a diff line is an insertion, defaults to \dQuote{> }}

\item{gutter.insert.ctd}{character(1L) if a diff line is wrapped, the
visual cue shifts to this character to indicate wrapping occured}

\item{gutter.delete}{character(1L) see \code{gutter.insert} above}

\item{gutter.delete.ctd}{character(1L) see \code{gutter.insert.ctd} above}

\item{gutter.match}{character(1L) see \code{gutter.insert} above}

\item{gutter.match.ctd}{character(1L) see \code{gutter.insert.ctd} above}

\item{gutter.guide}{character(1L) see \code{gutter.insert} above}

\item{gutter.guide.ctd}{character(1L) see \code{gutter.insert.ctd} above}

\item{gutter.fill}{character(1L) see \code{gutter.insert} above}

\item{gutter.fill.ctd}{character(1L) see \code{gutter.insert.ctd} above}

\item{gutter.pad}{character(1L) separator between gutter characters and the
rest of a line in a diff}

\item{pad.col}{character(1L) separator between columns in side by side mode}
}
\value{
a StyleText S4 object
}
\description{
Various character tokens are used throughout diffs to provide visual cues.
For example, gutters will contain characters that denote deletions and
insertions (\code{<} and \code{>} by default).
}
\seealso{
\code{\link{Style}}
}

